requirements.txt
pdfminer.six
Pillow